+++ packages/entity-graph-htmx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prometheus-ags/entity-graph-htmx
 
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @prometheus-ags/entity-graph-core@3.0.3
+  - @prometheus-ags/entity-graph-sdl@3.0.3
+
 ## 3.0.0
 
 ### Patch Changes