+++ website/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prometheus-ags/entity-management-docs
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @prometheus-ags/entity-graph-core@3.0.3
+  - @prometheus-ags/prometheus-entity-management@3.0.3
+
 ## 0.0.1
 
 ### Patch Changes